Changed name to indicate loading from an existing architecture
